pyproject.toml
[project]
name = "logcafe"
version = "0.1.0"
description = "TUI for viewing logs"
authors = [
    { name = "PokkaKiyo", email = "e0030925@u.nus.edu" }
]
dependencies = [
    "textual-dev>=1.6.1",
    "watchfiles>=0.24.0",
    "msgspec>=0.18.6",
    "rtoml>=0.11.0",
    "arrow>=1.3.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/logcafe"]
